One worker thread
