+++ apps/diegesis-server/config/debug_user_config.json
@@ -11,5 +11,5 @@
   "includeMutations": true,
   "staticPath": "../diegesis-user-client/build",
   "redirectToRoot": ["/who", "/how", "/list", "/entry/*"],
-  "deleteGenerated": false
+  "deleteGenerated": true
 }
