+++ apps/diegesis-server/src/lib/worker.js
@@ -1,0 +1,5 @@
+const {parentPort, workerData} = require("node:worker_threads");
+
+parentPort.on("message", data => {
+    postMessage(data);
+});
